Moving the repo from personal into IDIA
The old repo was hosted on my personal gitlab page, this commit moves
the development to the IDIA repo.

All the work for this commit was done by Joe Bochenek.

This commit contains the initial functions to create a binned UV plane,
potentially both standard deviation and median absolute deviation. It
also contains the singularity def file to build a container with all the
dependencies required to run this package.