Added function to compute UV from antenna coords
Parsing the min and max UV from the entire MS is very IO intensive and
slow. This computes the approximate maximum UV from the antenna
coordinates and source position.

If the outer edge of the UV plane is known to within few tens of bins,
it will not affect the efficacy of the flagging since most of the RFI is
at the smaller UV lengths.