Added RESIDUAL column support - fixed stokes selection
The Stokes selection now uses the real values of the Stokes parameters.
Added support to calculate the RESIDUAL column on the fly - which only
works when the CORRECTED_DATA and MODEL columns are present.

At the moment the code cascades to select DATA if none of RESIDUAL or
CORRECTED_DATA exists, but this will be turned into a command line
option at a later date.